+++ services/temp_tracker/src/prawnqtt.rs
@@ -1,7 +1,6 @@
 use super::model;
 use crossbeam_channel as channel;
-use rumqtt::{MqttCallback, MqttClient, MqttOptions, QoS};
-use uuid::Uuid;
+use rumqtt::MqttCallback;
 
 fn deser_message(msg: &rumqtt::Message) -> Option<model::TempMessage> {
     serde_json::from_str(std::str::from_utf8(&*msg.payload).unwrap())
@@ -9,13 +8,7 @@ fn deser_message(msg: &rumqtt::Message) -> Option<model::TempMessage> {
         .unwrap_or(None)
 }
 
-pub fn start_mqtt(
-    update_s: channel::Sender<model::TempMessage>,
-    mq_host: &str,
-    mq_port: u16,
-    mq_topic: &str,
-    mq_keep_alive: u16,
-) {
+pub fn create_mqtt_callback(update_s: channel::Sender<model::TempMessage>) -> MqttCallback {
     let deserialize_and_forward = move |msg: rumqtt::Message| {
         println!("Message on {:?}", msg.topic);
         let deser: Option<model::TempMessage> = deser_message(&msg);
@@ -34,21 +27,5 @@ pub fn start_mqtt(
         }
     };
 
-    let mq_message_callback = MqttCallback::new().on_message(deserialize_and_forward);
-
-    // Specify client connection options
-    let opts: MqttOptions = MqttOptions::new()
-        .set_keep_alive(mq_keep_alive)
-        .set_reconnect(3)
-        .set_client_id(generate_mq_client_id())
-        .set_broker(&format!("{}:{}", mq_host, mq_port)[..]);
-
-    MqttClient::start(opts, Some(mq_message_callback))
-        .expect("MQTT client couldn't start")
-        .subscribe(vec![(mq_topic, QoS::Level0)])
-        .unwrap()
-}
-
-fn generate_mq_client_id() -> String {
-    format!("sensor_tracker/{}", Uuid::new_v4())
+    MqttCallback::new().on_message(deserialize_and_forward)
 }
