+++ services/tracker_support/Cargo.toml
@@ -6,6 +6,8 @@ edition = "2018"
 
 [dependencies]
 envy = "0.3"
+rumqtt = "0.10"
 serde = "1.0"
 serde_derive = "1.0"
 serde_json = "1.0"
+uuid = { version = "0.7", features = ["v4", "v5"] }
