+++ services/tracker_support/src/lib.rs
@@ -1,7 +1,10 @@
-use envy;
-
+extern crate rumqtt;
 #[macro_use]
 extern crate serde_derive;
+extern crate uuid;
+
+use rumqtt::{MqttCallback, MqttClient, MqttOptions, QoS};
+use uuid::Uuid;
 
 #[derive(Deserialize, Debug, Clone)]
 pub struct TrackerConfig {
@@ -22,4 +25,28 @@ impl TrackerConfig {
             Err(e) => panic!("Unable to parse config ({})", e),
         }
     }
+}
+
+pub fn start_mqtt(
+    mq_message_callback: MqttCallback,
+    mq_host: &str,
+    mq_port: u16,
+    mq_topic: &str,
+    mq_keep_alive: u16,
+) {
+    // Specify client connection options
+    let opts: MqttOptions = MqttOptions::new()
+        .set_keep_alive(mq_keep_alive)
+        .set_reconnect(3)
+        .set_client_id(generate_mq_client_id())
+        .set_broker(&format!("{}:{}", mq_host, mq_port)[..]);
+
+    MqttClient::start(opts, Some(mq_message_callback))
+        .expect("MQTT client couldn't start")
+        .subscribe(vec![(mq_topic, QoS::Level0)])
+        .unwrap()
+}
+
+fn generate_mq_client_id() -> String {
+    format!("sensor_tracker/{}", Uuid::new_v4())
 }
